requirements.txt
Django
django-location-field
Pillow
djangorestframework

